[otbn] Factor loop stack out of state.py
No functional change; just breaking up a big file.
